+++ src/view/edit-point.js
@@ -2,8 +2,6 @@ import { nanoid } from 'nanoid';
 import { createElement } from '../utils/common';
 import { formatDate, getActualDate } from '../utils/date';
 import { Types, Destinations } from '../const';
-import { createPhotoContainerTemplate } from './pictures';
-import { createOffersContainerTemplate } from './offers';
 
 const BLANK_POINT = {
   type: Types.FLIGHT,
@@ -27,7 +25,6 @@ const BLANK_POINT = {
   isFavorite: false,
   id: nanoid(),
 };
-// console.log(BLANK_POINT);
 
 const createTypeItemTemplate = (eventType) => {
   const type = eventType.toLowerCase();
@@ -48,6 +45,38 @@ const createCheckedTypeItemTemplate = (eventType) => {
 };
 
 const createOptionTemplate = (point) => `<option value="${point}"></option>`;
+
+const createPhotoTemplate = ({src}) => `<img class="event__photo" src="${src}" alt="Event photo">`;
+
+const createOfferTemplate = ({isChecked, title, price}) => (
+  `<div class="event__offer-selector">
+      <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox" name="event-offer-comfort" ${isChecked ? 'checked' : ''}>
+      <label class="event__offer-label" for="event-offer-comfort-1">
+        <span class="event__offer-title">${title}</span>
+        &plus;&euro;&nbsp;
+        <span class="event__offer-price">${price}</span>
+      </label>
+    </div>`
+);
+
+const createOffersContainerTemplate = (offers) => (
+  offers ? `<section class="event__section  event__section--offers">
+    <h3 class="event__section-title  event__section-title--offers">Offers</h3>
+    <div class="event__available-offers">
+      ${offers.map(createOfferTemplate).join('')}
+    </div>
+  </section>` : ''
+);
+
+const createPhotoContainerTemplate = ({pictures}) => {
+  const photos = pictures.map(createPhotoTemplate).join('');
+
+  return `<div class="event__photos-container">
+    <div class="event__photos-tape">
+      ${photos}
+    </div>
+  </div>`;
+};
 
 const createEditPointTemplate = ({type, destination, dateFrom, dateTo, basePrice, offers}) => {
   const submenu = Object.values(Types).map((eventType) => eventType !== type
