--- src/view/offers.js
@@ -1,19 +1,0 @@
-const createOfferTemplate = ({isChecked, title, price}) => (
-  `<div class="event__offer-selector">
-      <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox" name="event-offer-comfort" ${isChecked ? 'checked' : ''}>
-      <label class="event__offer-label" for="event-offer-comfort-1">
-        <span class="event__offer-title">${title}</span>
-        &plus;&euro;&nbsp;
-        <span class="event__offer-price">${price}</span>
-      </label>
-    </div>`
-);
-
-export const createOffersContainerTemplate = (offers) => (
-  offers ? `<section class="event__section  event__section--offers">
-    <h3 class="event__section-title  event__section-title--offers">Offers</h3>
-    <div class="event__available-offers">
-      ${offers.map(createOfferTemplate).join('')}
-    </div>
-  </section>` : ''
-);
